Fix scrubbing not working.
In the previous checkin I had changed the setRequestedPlayheadPosition to take an additional parameter for the songID that we wanted to set the playhead position for.

This does not work because the requestedPlayheadPosition property is bound to the timelineBar.
In order for the binding to be able to call the property setter it must have only one parameter, the value, so adding more parameters broke it.

It is now changed back and working. I am wondering if binding is the best solution for keeping the slider knob and the song playback synchronized.