Add event chain to automate and branch off the chain of events that need to happen in the background whenever a new song is loaded.
When a new song url is passed to the app that hasn't been seen before it should:
1) get fingerprinted.
2) Exchange fingerprint for a uuid
3) look up metadata that is not available locally using uuid
4) Find album art using uuid

The idea is to get a chain of events where the success of a stage triggers the next without holding up the execution of the main thread.
Also, this should show the feasibility of using event notification in place of key-value binding in some cases where more objects need to simultaneously react to a change (eg. a selection of a sweet spot).