Disable some spammy comments.
I discovered the reason why there is a lag when fast scrolling the
songs.

I narrowed the lag down to a call to requestSongsFromAlbumWithName:
It is called in its own private thread so I couldn't understand why it
was causing lagging in the main thread where the UI was happening.

The problem is that, on the main thread, the song accessor `selectedSweetSpot`
is being called. All song.TEOData accessors that return a value go through
the TEOData managed context's thread using a performBlockAndWait.

At the same time a separate thread is calling
requestSongsFromAlbumWithName: which is performing a Core Data fetch
also on the TEOData managed Context. This takes some time and while this
is going on the main thread is waiting for the performBlockAndWait to
stop waiting and return, hence the lag.

The solution is not immediately obvious (or rather the obvious solution is not
immediately appealing) since it involves going back to what I was doing before
which was to keep a separate dictionary where I tracked modified songs that
would need saving.
I'm hoping a better solution comes to mind that allows me to keep the changeable data
as managed objects that implicitly know when they need saving.
On the other hand, since I have a bit of a mess with two Core Data
stores, one for the songUserData and one for the TEOSongData, and I'd
like to unify it, I probably need to change the whole save doodad
anyway.