Make song property access fast.
Moved song properties previously grouped as a managed object TEOData
to be properties of the song itself so their access is unencumbered by
having to go through a the same thread as the managed object context.

This has also meant changes need to be made to how the songs that need
saving will be handled. This is stil work in progress.

Concurrently I'm still trying to find and reduce the clear lag that
occurs when scrolling quickly past a group of songs whose covers have
been revealed and which are animated as the cursor moves over them.

Firstly I need to ensure that there is nothing dumb going on like
re-fetching the images every time and then look at whether I'm doing the
display part as efficiently as possible.