Add swizzled category on NSImage.
In order to avoid setting a cell's image to something it is already set
to and to avoid filling the cover image cache up with duplicates I must
be able to compare two images.

This, I have done using an MD5 hashing function that is called
asynchronously on object creation.
To avoid having to recalculate the hash I store it with the image
itself. By making a swizzled category on NSImage I can add an id property that can be
looked up immediately once calculated and used for comparing two such
objects.

I cannot use just a category as they don't allow instance variables.

I cannot use a class extension as they require the source of the
extended class be available. Framework classes like NSImage are not.

I cannot just add an associated value
(http://nshipster.com/associated-objects/) to NSImage and then make a
category of it, because I can still not override the imageNamed
convenience init and call the original from inside it. I need to do that
to get all the searching in xibs for the image. I suppose I could avoid
using the imageNamed convenience method altogether but, well it's
convenient. Besides it's a challenge now.

I don't want to use subclassing as I'd have to change all the calls to
use my subclass instead and it also couples all the methods that use it
too tightly.

So I end up using a category so that I can continue to use the NSImage inits (and
the convenience init imageNamed) without having to change any calls. But
to be able to call the original NSImage imageNamed:
I have to swizzle the method with my own that calls the original and then do my own
hashing thing.