Fix audio loading code to wait for ready status.
The code so far has not been properly waiting for the SongPlayerItem
status to turn ready, probably causing some weirdness.
By adding an observer on the SongPlayerItem's status keypath and putting
the desired callback method into a block (which was needed in order to
capture the startTime passed to the playback method) we now correctly
wait for a ready status and only then call the callback.