Simplify cache and playback flow.
Move init stuff from awakeFromNib, which was being called multiple
times, into viewDidLoad. This fixed the observer being added twice and
subsequently firing twice.

Change requestImageForSongID to call the passed image handler block with
the default No Cover image instead of nil when failing to find a cover
image anywhere.

Add a property to the SongPool to keep track of the No Cover image hash
id so that it can easily be retrived from the cover art local cache.