Fix cover animation issue.
It turns out that the songCell matrix needed to be told to update
explicitly (which was why the TGGridCell drawWithFrame was only being
called at the inital creation of the grid).

The solution was to set a grid cell to be flagged as hidden at the start
of a layer animation and set it back to not hidden at animation
completion. To ensure that drawWithFrame got called after changing a
cell's image (or its hidden status) I added setNeedsDisplayInRect calls
after such a change. Note the InRect part which optimizes the refresh to
only redraw the updated cell.