Add two new WIP labels: CACH and COVR.
The two represent work in progress on:

1) caching which I want to have working without state,
 which means the old cache should be an input and
the new cache an output (or in this case an input to a handler that the
caching method takes as an argument).

2) The cover which I have added a songPoolStartedFetchingSongID delegate
method for so that the songpool can cause the _fetching_ cover to be
applied to a song immediately, and not have to wait for the song to have
started playing. This is currently not working because songs that have
been marked as selected are not necessarily successfully cached because
the caching may cancel in-flight and thus not call the cover refresh method for
the song.