Fix caching.
Made the caching work by creating a cache queue and a callback queue.

The cache creation method newCacheFromCache: takes a previous cache as an input to avoid
having to reference a global cache and to ensure that it stays in sync
with reality (ie whether the actual songs are loaded).

To ensure that newCacheFromCache: always has a previous cache to work
with I have wrapped its calls in an async method
, performHandlerWhenCacheIsReady:, that will either pick
an available cache from a cache queue and call the callback with it or,
if none is available, to add its callback (which is the call to newCacheFromCache:)
to the callback queue.

Inside newCacheFromCache: when a new cache has been made it will see if
there are any callbacks waiting in the callback queue.
If there are it will dequeue one and call it with the new cache.
If there are none it will simply put the new cache onto the cache queue
ready for the next call to performHandlerWhenCacheIsReady:.

Because newCacheFromCache: only adds its newly created cache to the
cache queue if there are no callbacks in the callback queue we can be
sure that, in performHandlerWhenCacheIsReady:, if the cache queue is
empty it must be because a new cache is being made and so we can simply
 add the callback to the callback queue knowing that it will get called
 inside the newCacheFromCache: once it is done.