The caching saga continues.
Having realized that my new, faster approach suffered from the issue
that there was the possibility of concurrently loading and unloading
songs, I have implemented serialized versions of the load and unload
methods (used to be named prepareForPlayback and cleanCache)
which should fix that particular problem. I am still seeing
songs being cached when they shouldn't and vice versa, so there's
clearly still something not quite right, but I still see the concept as
basically sound.

The concept, to re-iterate is this:

Whenever a song is explicitly selected it is loaded and its id added to
a set, selectedSongsCache, to keep track of it.

Immediately after we cancel any currenly running caching operations
(initiated by previous calls to newCacheFromContext) and then we call
newCacheFromContext with the given context.

If a running caching operation is cancelled it returns the ids of the
songs it managed to load before the cancellation.
These are unloaded immediately by the operation's completion block.

If a caching operation is successful it will return the new cache.
The operation's completion block will unload the songs in the old
cache that are not in the new cache and will also unload the songs in
the selectedSongsCache that are not in the new cache.