Add cell frame relative UI positioning.
Have doubts as to whether the UI should be centered on the cell or
centered on the mouse pointer at click time. The latter seems more
dynamic, especially if it animates nicely into position.
If I choose that approach I'll need to disable the selection of other
songs whilst the UI is up.