Get split view panes toggling properly.
The effect is simply (but inscrutably) to set the holding priority of
the middle split view pane of the split view controller to 1000 and the
other two to 250. No other ratios I've tried worked. Eg.

    249, 250, 249 nope
    250, 260, 250 no joy
    250, 750, 250 arse

The solution is not complete because I can now, as a consequence of the
1000 holding priority of the middle view, not resize the cover view
horizontally. Also, there seems to be a limiting factor of the screen
size as the panels refuse to go beyond the screen width and will
push the unpushable (1000 holding priority, remember?) middle pane
instead.